Remove outdated migration files
Removed the migration files `20250203191330_AddedType.Designer.cs` and `20250203191330_AddedType.cs`, which added a `Type` column to the `Products` table and updated seed data for the `Memberships` table. These changes will no longer be applied.

Also removed the `20250203195157_InitialCreate.Designer.cs` and `20250203195157_InitialCreate.cs` files, which contained the initial database schema creation logic for various entities in the `Bliss` namespace. This includes the creation of multiple tables and initial data insertion for the `Memberships` table. The removal suggests a shift to a different schema or approach.